# openapi.yml
openapi: 3.1.0
info:
  title: Company Database Connector
  description: >
    Allows GPT to run safe, read-only SQL queries on the company and finance
    schemas of a PostgreSQL database via authenticated endpoints.
  version: "1.0.0"
servers:
  - url: https://e7f4f5645351.ngrok-free.app

paths:
  /user/query:
    post:
      operationId: queryCompanyDB
      summary: Run a read-only SQL query as a standard user
      description: >
        Executes a SELECT query on the company schema in PostgreSQL using a
        read-only user role. Requires a valid USER_API_KEY bearer token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sql:
                  type: string
                  description: SQL SELECT query to execute (must reference the company schema).
                  example: "SELECT first_name, last_name FROM company.employees LIMIT 3;"
      responses:
        "200":
          description: Successful query
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  rows:
                    type: integer
                  results:
                    type: array
                    items:
                      type: object
                      additionalProperties: true
                  role:
                    type: string
                    example: user
        "400":
          description: Invalid SQL or server error
        "403":
          description: Invalid or missing user API key

  /admin/query:
    post:
      operationId: queryAdminDB
      summary: Run a read-only SQL query as an admin
      description: >
        Executes a SELECT query on the company or finance schemas in PostgreSQL
        using the admin role. Requires a valid ADMIN_API_KEY bearer token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sql:
                  type: string
                  description: >
                    SQL SELECT query to execute (can reference company or finance schemas).
                  example: "SELECT * FROM finance.transactions LIMIT 5;"
      responses:
        "200":
          description: Successful query
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  rows:
                    type: integer
                  results:
                    type: array
                    items:
                      type: object
                      additionalProperties: true
                  role:
                    type: string
                    example: admin
        "400":
          description: Invalid SQL or server error
        "403":
          description: Invalid or missing admin API key